Add VideoScript Model
VideoScripts are scripts that will reside in a directory inside
the library folder. All videos in that directory, as well as in
subdirectories, will have this video script executed, before the
video is played. Useful for a generic solution to pre-processing
videos before playing.

This commit adds the VideoScript model. Crucially there is a many
to many relationship between scripts and their videos.